refactor: AuthDialog를 로그인과 회원가입 로직을 분리 #33
