+++ src/apps/home/components/auth/index.ts
@@ -1,2 +1,5 @@
 export { AuthDialog } from './AuthDialog';
+export { LoginForm } from './LoginForm';
+export { SignupForm } from './SignupForm';
+export { authSchema, type AuthFormData } from './schemas';
 export { openLoginDialog, openSignupDialog } from './authHelpers';
